+++ resources/assets/less/main.less
@@ -1,0 +1,108 @@
+body {
+    font-family: 'Roboto', sans-serif;
+}
+#map { height: 100%;width:100%;position:fixed;top:0;left:0; }
+
+.leaflet-popup {
+    margin-bottom:40px;
+}
+
+.btn {
+  background-color:#F1301A;
+  border:2px solid #F1301A;
+}
+
+.alert-success {
+    background-color: #dff0d8;
+    border-color: #d6e9c6;
+    color: #3c763d;
+    font-size: 20px !important;
+    margin:0px auto;
+    margin-top: 50px;
+    position: relative;
+    z-index: 99999;
+    width:70%;
+}
+
+.alert-success h2 {
+  font-size:18px;
+}
+
+.nav {
+    height:80px;
+    background-color:rgba(0,0,0,.8);
+    width:100%;
+    position:fixed;
+    top:0;
+    left:0;
+    z-index:99999;
+    color:#fff;
+    font-weight:300;
+}
+
+.nav h2 {
+    font-weight:300;
+}
+
+.nav .btn {
+    margin-top:15px;
+    padding:10px 30px;
+}
+
+.leaflet-left {
+    left: 0;
+    top: 90px;
+}
+
+.jake {
+  background-color:#F1341E;
+  padding:10px 20px;
+  color:#fff;
+  position:fixed;
+  bottom:0;
+  left:0;
+
+  &:hover {
+    background-color:darken(#F1341E,15);
+    text-decoration:none;
+    color:#fff;
+  }
+}
+
+.fa-heart {
+  color:#F1341E;
+}
+
+
+
+.modal {
+    z-index:9999999;
+}
+
+.nav { 
+
+  h2 {
+      margin-top:20px;
+  }
+
+  h3 {
+    font-weight:300;
+    margin-top:23px;
+    text-align: center;
+  }
+
+}
+
+.logo span {
+  font-weight:600;
+  font-size:33px;
+}
+
+.alert-danger {
+    background-color: #f2dede;
+    border-color: #ebccd1;
+    color: #a94442;
+    position: relative;
+    z-index: 9999;
+    margin-top:100px;
+}
